requirements.txt
# Flask Core
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.27

# Environment Management
python-dotenv==1.0.1

# Validation Libraries
getdc==0.2.0
inspectorfhir==0.1.0

# Rate Limiting
Flask-Limiter==3.5.1

# Testing
pytest==8.0.2
pytest-flask==1.3.0
pytest-cov==4.1.0

# Utilities
requests==2.31.0
